refactor: member card 파일 이동
